added action_from_check to ruff_utils, moved state macros to ruffd_types
