+++ ruffd-core/src/ruff_utils.rs
@@ -1,5 +1,7 @@
 use ruffd_types::lsp_types;
 use ruffd_types::ruff::checks::Check;
+use std::collections::HashMap;
+
 pub fn diagnostic_from_check(check: &Check) -> lsp_types::Diagnostic {
     let range = {
         // diagnostic is zero indexed, but message rows are 1-indexed
@@ -33,4 +35,41 @@ pub fn diagnostic_from_check(check: &Check) -> lsp_types::Diagnostic {
         related_information: None,
         data: None,
     }
+}
+
+pub fn action_from_check(
+    check: &Check,
+    document_uri: &lsp_types::Url,
+) -> Option<lsp_types::CodeAction> {
+    check.fix.as_ref().map(|fix| {
+        let row_start = fix.patch.location.row() as u32 - 1;
+        let row_end = fix.patch.end_location.row() as u32 - 1;
+        let col_start = fix.patch.location.column() as u32;
+        let col_end = fix.patch.end_location.column() as u32;
+        lsp_types::CodeAction {
+            title: format!("fix {}", check.kind.code().as_ref()),
+            kind: Some(lsp_types::CodeActionKind::QUICKFIX),
+            diagnostics: Some(vec![diagnostic_from_check(check)]),
+            edit: Some(lsp_types::WorkspaceEdit {
+                changes: Some(HashMap::from_iter(vec![(
+                    document_uri.clone(),
+                    vec![lsp_types::TextEdit {
+                        range: lsp_types::Range {
+                            start: lsp_types::Position {
+                                line: row_start,
+                                character: col_start,
+                            },
+                            end: lsp_types::Position {
+                                line: row_end,
+                                character: col_end,
+                            },
+                        },
+                        new_text: fix.patch.content.clone(),
+                    }],
+                )])),
+                ..Default::default()
+            }),
+            ..Default::default()
+        }
+    })
 }
